// src/shared/libs/types.ts
export type TAuction = {
    id: number
    image: string
    altImage: string
    avatar: string
    avatarAlt: string
    name: string
    likes: number
    timer: string
    bid: string
    isLiked: boolean
}

export type TCollection = {
    image: string
    altImage: string
    avatar: string
    avatarAlt: string
    name: string
    number: string

}

export type TCreator = {
    avatar: string
    avatarAlt: string
    name: string
    money: string
    number: number
}